server/requirements.txt
Django==2.0.9
django-cleanup==2.1.0
django-cors-headers==2.2.0
djangorestframework==3.8.2
django-webpack-loader==0.6.0
channels==2.1.2
channels_redis==2.2.1
celery==4.2.0
whitenoise==3.3.1
pytz==2018.4
python-magic==0.4.15
flake8
tox
psycopg2
spacy
